CLOUD-476: Rename Process Execution Server to Intelligent Process Server
