+++ processserver/library-client/src/main/webapp/WEB-INF/web.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<web-app>
+  <servlet>
+    <servlet-name>LibraryServlet</servlet-name>
+    <servlet-class>org.openshift.quickstarts.processserver.library.client.LibraryServlet</servlet-class>
+  </servlet>
+  <servlet-mapping>
+    <servlet-name>LibraryServlet</servlet-name>
+    <url-pattern>/*</url-pattern>
+  </servlet-mapping>
+</web-app>
