+++ processserver/library/src/main/java/org/openshift/quickstarts/processserver/library/types/Loan.java
@@ -1,0 +1,135 @@
+//
+// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.5 
+// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
+// Any modifications to this file will be lost upon recompilation of the source schema. 
+// Generated on: 2014.09.04 at 12:38:20 PM EDT 
+//
+
+
+package org.openshift.quickstarts.processserver.library.types;
+
+import java.io.Serializable;
+
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlElement;
+import javax.xml.bind.annotation.XmlType;
+
+
+/**
+ * <p>Java class for loan complex type.
+ * 
+ * <p>The following schema fragment specifies the expected content contained within this class.
+ * 
+ * <pre>
+ * &lt;complexType name="loan">
+ *   &lt;complexContent>
+ *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *       &lt;sequence>
+ *         &lt;element name="id" type="{http://www.w3.org/2001/XMLSchema}string" minOccurs="0"/>
+ *         &lt;element name="approved" type="{http://www.w3.org/2001/XMLSchema}boolean"/>
+ *         &lt;element name="notes" type="{http://www.w3.org/2001/XMLSchema}string" minOccurs="0"/>
+ *         &lt;element ref="{urn:openshift-quickstarts:processserver-library:1.3}book" minOccurs="0"/>
+ *       &lt;/sequence>
+ *     &lt;/restriction>
+ *   &lt;/complexContent>
+ * &lt;/complexType>
+ * </pre>
+ * 
+ * 
+ */
+@SuppressWarnings("serial")
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlType(name = "loan", namespace = "urn:openshift-quickstarts:processserver-library:1.3", propOrder = {
+    "id",
+    "approved",
+    "notes",
+    "book"
+})
+public class Loan implements Serializable {
+
+    protected long id;
+    protected boolean approved;
+    protected String notes;
+    @XmlElement(namespace = "urn:openshift-quickstarts:processserver-library:1.3")
+    protected Book book;
+
+    /**
+     * Gets the value of the id property.
+     */
+    public long getId() {
+        return id;
+    }
+
+    /**
+     * Sets the value of the id property.
+     */
+    public void setId(long value) {
+        this.id = value;
+    }
+
+    /**
+     * Gets the value of the approved property.
+     * 
+     */
+    public boolean isApproved() {
+        return approved;
+    }
+
+    /**
+     * Sets the value of the approved property.
+     * 
+     */
+    public void setApproved(boolean value) {
+        this.approved = value;
+    }
+
+    /**
+     * Gets the value of the notes property.
+     * 
+     * @return
+     *     possible object is
+     *     {@link String }
+     *     
+     */
+    public String getNotes() {
+        return notes;
+    }
+
+    /**
+     * Sets the value of the notes property.
+     * 
+     * @param value
+     *     allowed object is
+     *     {@link String }
+     *     
+     */
+    public void setNotes(String value) {
+        this.notes = value;
+    }
+
+    /**
+     * Gets the value of the book property.
+     * 
+     * @return
+     *     possible object is
+     *     {@link Book }
+     *     
+     */
+    public Book getBook() {
+        return book;
+    }
+
+    /**
+     * Sets the value of the book property.
+     * 
+     * @param value
+     *     allowed object is
+     *     {@link Book }
+     *     
+     */
+    public void setBook(Book value) {
+        this.book = value;
+    }
+
+}
