+++ processserver/library/src/main/java/org/openshift/quickstarts/processserver/library/types/LoanResponse.java
@@ -1,0 +1,74 @@
+//
+// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.5 
+// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
+// Any modifications to this file will be lost upon recompilation of the source schema. 
+// Generated on: 2014.09.04 at 12:38:20 PM EDT 
+//
+
+
+package org.openshift.quickstarts.processserver.library.types;
+
+import java.io.Serializable;
+
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlElement;
+import javax.xml.bind.annotation.XmlRootElement;
+import javax.xml.bind.annotation.XmlType;
+
+
+/**
+ * <p>Java class for anonymous complex type.
+ * 
+ * <p>The following schema fragment specifies the expected content contained within this class.
+ * 
+ * <pre>
+ * &lt;complexType>
+ *   &lt;complexContent>
+ *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *       &lt;sequence>
+ *         &lt;element ref="{urn:openshift-quickstarts:processserver-library:1.3}loan"/>
+ *       &lt;/sequence>
+ *     &lt;/restriction>
+ *   &lt;/complexContent>
+ * &lt;/complexType>
+ * </pre>
+ * 
+ * 
+ */
+@SuppressWarnings("serial")
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlType(name = "loanResponse", namespace = "urn:openshift-quickstarts:processserver-library:1.3", propOrder = {
+    "loan"
+})
+@XmlRootElement(name = "loanResponse", namespace = "urn:openshift-quickstarts:processserver-library:1.3")
+public class LoanResponse implements Serializable {
+
+    @XmlElement(namespace = "urn:openshift-quickstarts:processserver-library:1.3", required = true)
+    protected Loan loan;
+
+    /**
+     * Gets the value of the loan property.
+     * 
+     * @return
+     *     possible object is
+     *     {@link Loan }
+     *     
+     */
+    public Loan getLoan() {
+        return loan;
+    }
+
+    /**
+     * Sets the value of the loan property.
+     * 
+     * @param value
+     *     allowed object is
+     *     {@link Loan }
+     *     
+     */
+    public void setLoan(Loan value) {
+        this.loan = value;
+    }
+
+}
