+++ processserver/library/src/main/resources/META-INF/kmodule.xml
