+++ processserver/pom.xml
@@ -1,0 +1,26 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+   <modelVersion>4.0.0</modelVersion>
+
+   <groupId>org.openshift.quickstarts</groupId>
+   <artifactId>processserver-parent</artifactId>
+   <version>1.3.0-SNAPSHOT</version>
+   <packaging>pom</packaging>
+
+   <name>OpenShift Quickstarts: Intelligent Process Server: Parent</name>
+   <description>OpenShift Quickstarts: Intelligent Process Server: Parent</description>
+
+    <licenses>
+        <license>
+            <name>Apache License, Version 2.0</name>
+            <distribution>repo</distribution>
+            <url>http://www.apache.org/licenses/LICENSE-2.0.html</url>
+        </license>
+    </licenses>
+
+    <modules>
+        <module>library</module>
+        <module>library-client</module>
+    </modules>
+
+</project>
