--- procexecserver/library-client/pom.xml
@@ -1,160 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-   <modelVersion>4.0.0</modelVersion>
-
-   <groupId>org.openshift.quickstarts</groupId>
-   <artifactId>procexecserver-library-client</artifactId>
-   <version>1.3.0-SNAPSHOT</version>
-   <packaging>war</packaging>
-
-   <name>OpenShift Quickstarts: Decision Server: Library - Client</name>
-   <description>A simple example application demonstrating the use of a business process and rules.</description>
-
-    <licenses>
-        <license>
-            <name>Apache License, Version 2.0</name>
-            <distribution>repo</distribution>
-            <url>http://www.apache.org/licenses/LICENSE-2.0.html</url>
-        </license>
-    </licenses>
-
-    <properties>
-        <!-- Explicitly declaring the source encoding eliminates the following message: -->
-        <!-- [WARNING] Using platform encoding (UTF-8 actually) to copy filtered resources, i.e. build is platform dependent! -->
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <!-- maven-compiler-plugin -->
-        <maven.compiler.target>1.8</maven.compiler.target>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <!-- NOTE: the "*-redhat-7" version can temporarily only be accessed by
-             downloading the zip version of that maven release's repository -->
-        <kie.version>6.3.0.Final-redhat-7</kie.version>
-    </properties>
-
-   <dependencies>
-      <dependency>
-         <groupId>org.jboss.spec.javax.servlet</groupId>
-         <artifactId>jboss-servlet-api_3.0_spec</artifactId>
-         <version>1.0.2.Final-redhat-2</version>
-         <scope>provided</scope>
-      </dependency>
-      <dependency>
-         <groupId>junit</groupId>
-         <artifactId>junit</artifactId>
-         <version>4.12</version>
-         <scope>test</scope>
-      </dependency>
-      <dependency>
-         <groupId>org.apache.activemq</groupId>
-         <artifactId>activemq-client</artifactId>
-         <version>5.11.1</version>
-         <scope>compile</scope>
-      </dependency>
-      <dependency>
-         <groupId>org.drools</groupId>
-         <artifactId>drools-compiler</artifactId>
-         <version>${kie.version}</version>
-         <scope>compile</scope>
-      </dependency>
-      <dependency>
-         <groupId>org.hornetq</groupId>
-         <artifactId>hornetq-jms-client</artifactId>
-         <version>2.3.25.SP3-redhat-1</version>
-         <scope>provided</scope>
-      </dependency>
-      <dependency>
-         <groupId>org.jboss</groupId>
-         <artifactId>jboss-remote-naming</artifactId>
-         <version>1.0.11.Final-redhat-1</version>
-         <scope>provided</scope>
-      </dependency>
-      <dependency>
-         <groupId>org.jboss.xnio</groupId>
-         <artifactId>xnio-nio</artifactId>
-         <version>3.0.14.GA-redhat-1</version>
-         <scope>provided</scope>
-      </dependency>
-      <dependency>
-         <groupId>org.jbpm</groupId>
-         <artifactId>jbpm-bpmn2</artifactId>
-         <version>${kie.version}</version>
-         <scope>compile</scope>
-      </dependency>
-      <dependency>
-         <groupId>org.kie.server</groupId>
-         <artifactId>kie-server-client</artifactId>
-         <version>${kie.version}</version>
-         <scope>compile</scope>
-      </dependency>
-      <!-- for remote debugging of MDB
-      <dependency>
-         <groupId>org.kie.server</groupId>
-         <artifactId>kie-server-jms</artifactId>
-         <version>${kie.version}</version>
-         <scope>compile</scope>
-      </dependency>
-      -->
-      <dependency>
-         <groupId>org.openshift.quickstarts</groupId>
-         <artifactId>procexecserver-library</artifactId>
-         <version>${project.version}</version>
-      </dependency>
-      <dependency>
-         <groupId>org.slf4j</groupId>
-         <artifactId>slf4j-simple</artifactId>
-         <version>1.7.2</version>
-         <scope>provided</scope>
-      </dependency>
-   </dependencies>
-
-   <build>
-      <plugins>
-         <plugin>
-            <groupId>org.codehaus.mojo</groupId>
-            <artifactId>exec-maven-plugin</artifactId>
-            <version>1.4.0</version>
-            <executions>
-               <execution>
-                  <goals>
-                     <goal>exec</goal>
-                  </goals>
-               </execution>
-            </executions>
-            <configuration>
-               <mainClass>org.openshift.quickstarts.procexecserver.library.client.LibraryClient</mainClass>
-            </configuration>
-         </plugin>
-      </plugins>
-   </build>
-
-    <repositories>
-        <repository>
-            <id>jboss-enterprise-techpreview-group</id>
-            <name>JBoss Enterprise Technology Preview  Maven Repository Group</name>
-            <url>http://maven.repository.redhat.com/techpreview/all/</url>
-            <layout>default</layout>
-            <releases>
-                <enabled>true</enabled>
-                <updatePolicy>never</updatePolicy>
-            </releases>
-            <snapshots>
-                <enabled>true</enabled>
-                <updatePolicy>never</updatePolicy>
-            </snapshots>
-        </repository>
-        <repository>
-            <id>jboss-public-repository-group</id>
-            <name>JBoss Public Maven Repository Group</name>
-            <url>https://repository.jboss.org/nexus/content/groups/public/</url>
-            <layout>default</layout>
-            <releases>
-                <enabled>true</enabled>
-                <updatePolicy>never</updatePolicy>
-            </releases>
-            <snapshots>
-                <enabled>true</enabled>
-                <updatePolicy>never</updatePolicy>
-            </snapshots>
-        </repository>
-    </repositories>
-
-</project>
