--- procexecserver/library/src/main/java/org/openshift/quickstarts/procexecserver/library/types/ReturnResponse.java
@@ -1,64 +1,0 @@
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.5 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2014.09.04 at 12:38:20 PM EDT 
-//
-
-
-package org.openshift.quickstarts.procexecserver.library.types;
-
-import java.io.Serializable;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlRootElement;
-import javax.xml.bind.annotation.XmlType;
-
-
-/**
- * <p>Java class for anonymous complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>
- * &lt;complexType>
- *   &lt;complexContent>
- *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       &lt;sequence>
- *         &lt;element name="acknowledged" type="{http://www.w3.org/2001/XMLSchema}boolean"/>
- *       &lt;/sequence>
- *     &lt;/restriction>
- *   &lt;/complexContent>
- * &lt;/complexType>
- * </pre>
- * 
- * 
- */
-@SuppressWarnings("serial")
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "returnResponse", namespace = "urn:openshift-quickstarts:procexecserver-library:1.0", propOrder = {
-    "acknowledged"
-})
-@XmlRootElement(name = "returnResponse", namespace = "urn:openshift-quickstarts:procexecserver-library:1.0")
-public class ReturnResponse implements Serializable {
-
-    protected boolean acknowledged;
-
-    /**
-     * Gets the value of the acknowledged property.
-     * 
-     */
-    public boolean isAcknowledged() {
-        return acknowledged;
-    }
-
-    /**
-     * Sets the value of the acknowledged property.
-     * 
-     */
-    public void setAcknowledged(boolean value) {
-        this.acknowledged = value;
-    }
-
-}
