--- procexecserver/library/src/main/java/org/openshift/quickstarts/procexecserver/library/types/Suggestion.java
@@ -1,72 +1,0 @@
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.5 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2014.09.04 at 12:38:20 PM EDT 
-//
-
-
-package org.openshift.quickstarts.procexecserver.library.types;
-
-import java.io.Serializable;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlType;
-
-
-/**
- * <p>Java class for suggestion complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>
- * &lt;complexType name="suggestion">
- *   &lt;complexContent>
- *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       &lt;sequence>
- *         &lt;element ref="{urn:openshift-quickstarts:procexecserver-library:1.0}book" minOccurs="0"/>
- *       &lt;/sequence>
- *     &lt;/restriction>
- *   &lt;/complexContent>
- * &lt;/complexType>
- * </pre>
- * 
- * 
- */
-@SuppressWarnings("serial")
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "suggestion", namespace = "urn:openshift-quickstarts:procexecserver-library:1.0", propOrder = {
-    "book"
-})
-public class Suggestion implements Serializable {
-
-    @XmlElement(namespace = "urn:openshift-quickstarts:procexecserver-library:1.0")
-    protected Book book;
-
-    /**
-     * Gets the value of the book property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Book }
-     *     
-     */
-    public Book getBook() {
-        return book;
-    }
-
-    /**
-     * Sets the value of the book property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Book }
-     *     
-     */
-    public void setBook(Book value) {
-        this.book = value;
-    }
-
-}
