--- procexecserver/library/src/main/resources/META-INF/LibraryTypes.xsd
@@ -1,86 +1,0 @@
-<xsd:schema
-  xmlns:xsd="http://www.w3.org/2001/XMLSchema"
-  xmlns:urn="urn:openshift-quickstarts:procexecserver-library:1.0"
-  targetNamespace="urn:openshift-quickstarts:procexecserver-library:1.0">
-
-  <xsd:element name="book" type="urn:book"/>
-  <xsd:complexType name="book">
-    <xsd:sequence>
-      <xsd:element name="isbn" type="xsd:string" minOccurs="0" maxOccurs="1"/>
-      <xsd:element name="title" type="xsd:string" minOccurs="0" maxOccurs="1"/>
-      <xsd:element name="synopsis" type="xsd:string" minOccurs="0" maxOccurs="1"/>
-    </xsd:sequence>
-  </xsd:complexType>
-
-  <xsd:element name="suggestion" type="urn:suggestion"/>
-  <xsd:complexType name="suggestion">
-    <xsd:sequence>
-      <xsd:element ref="urn:book" minOccurs="0" maxOccurs="1"/>
-    </xsd:sequence>
-  </xsd:complexType>
-
-  <xsd:element name="loan" type="urn:loan"/>
-  <xsd:complexType name="loan">
-    <xsd:sequence>
-      <xsd:element name="id" type="xsd:long" minOccurs="0" maxOccurs="1"/>
-      <xsd:element name="approved" type="xsd:boolean" minOccurs="0" maxOccurs="1"/>
-      <xsd:element name="notes" type="xsd:string" minOccurs="0" maxOccurs="1"/>
-      <xsd:element ref="urn:book" minOccurs="0" maxOccurs="1"/>
-    </xsd:sequence>
-  </xsd:complexType>
-
-  <!-- XMLRootElement name is 'getSuggestion', but name for type generation is 'suggestionRequest' -->
-  <xsd:element name="suggestionRequest">
-    <xsd:complexType>
-      <xsd:sequence>
-        <xsd:element name="keyword" type="xsd:string" minOccurs="1" maxOccurs="1"/>
-      </xsd:sequence>
-    </xsd:complexType>
-  </xsd:element>
-
-  <!-- XMLRootElement name is 'getSuggestionResponse', but name for type generation is 'suggestionResponse' -->
-  <xsd:element name="suggestionResponse">
-    <xsd:complexType>
-      <xsd:sequence>
-        <xsd:element ref="urn:suggestion" minOccurs="0" maxOccurs="1"/>
-      </xsd:sequence>
-    </xsd:complexType>
-  </xsd:element>
-
-  <!-- XMLRootElement name is 'attemptLoan', but name for type generation is 'loanRequest' -->
-  <xsd:element name="loanRequest">
-    <xsd:complexType>
-      <xsd:sequence>
-        <xsd:element name="isbn" type="xsd:string" minOccurs="1" maxOccurs="1"/>
-      </xsd:sequence>
-    </xsd:complexType>
-  </xsd:element>
-
-  <!-- XMLRootElement name is 'attemptLoanResponse', but name for type generation is 'loanResponse' -->
-  <xsd:element name="loanResponse">
-    <xsd:complexType>
-      <xsd:sequence>
-        <xsd:element ref="urn:loan" minOccurs="1" maxOccurs="1"/>
-      </xsd:sequence>
-    </xsd:complexType>
-  </xsd:element>
-
-  <!-- XMLRootElement name is 'returnLoan', but name for type generation is 'returnRequest' -->
-  <xsd:element name="returnRequest">
-    <xsd:complexType>
-      <xsd:sequence>
-        <xsd:element ref="urn:loan" minOccurs="1" maxOccurs="1"/>
-      </xsd:sequence>
-    </xsd:complexType>
-  </xsd:element>
-
-  <!-- XMLRootElement name is 'returnLoanResponse', but name for type generation is 'returnResponse' -->
-  <xsd:element name="returnResponse">
-    <xsd:complexType>
-      <xsd:sequence>
-        <xsd:element name="acknowledged" type="xsd:boolean" minOccurs="1" maxOccurs="1"/>
-      </xsd:sequence>
-    </xsd:complexType>
-  </xsd:element>
-
-</xsd:schema>
